Refactor todo service
