+++ src/services/todo/todo-api-types.d.ts
@@ -1,0 +1,7 @@
+export type TodoId = string;
+
+export type Todo = {
+  id: TodoId;
+  content: string;
+  completed: boolean;
+};
